movies_admin/requirements.txt
django==3.2
flake8==4.0
python-dotenv==0.21.0
django-split-settings==1.2.0